Sprint7

+++ Luqmit3ish/Luqmit3ish/Services/FoodServices.cs
@@ -1,7 +1,9 @@
 using Luqmit3ish.Connection;
 using Luqmit3ish.Exceptions;
+using Luqmit3ish.Interfaces;
 using Luqmit3ish.Models;
 using Luqmit3ish.Utilities;
+using Luqmit3ish.ViewModels;
 using Newtonsoft.Json;
 using System;
 using System.Collections.ObjectModel;
@@ -12,10 +14,11 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
+using Xamarin.Essentials;
 
 namespace Luqmit3ish.Services
 {
-    class FoodServices
+    class FoodServices : IFoodServices
     {
         private readonly HttpClient _httpClient;
         private readonly string _apiUrl = Constants.BaseUrl + "api/Food";
@@ -24,7 +27,7 @@ namespace Luqmit3ish.Services
         public FoodServices()
         {
             _httpClient = new HttpClient();
-            _connection = new Connection();
+            _connection = new InternetConnection();
         }
 
         public async Task<ObservableCollection<Dish>> GetFood()
@@ -126,7 +129,7 @@ namespace Luqmit3ish.Services
             }
         }
 
-        public async Task<bool> UpdateDish(DishRequest dishRequest, int food_id)
+        public async Task<bool> UpdateDish(Dish dishRequest)
         {
             if (!_connection.CheckInternetConnection())
             {
@@ -134,9 +137,20 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var json = JsonConvert.SerializeObject(dishRequest);
                 var content = new StringContent(json, Encoding.UTF8, "application/json");
-                var response = await _httpClient.PutAsync(_apiUrl + "/" + food_id, content);
+                var response = await _httpClient.PutAsync(_apiUrl + "/" + dishRequest.Id, content);
+                await UploadPhoto(dishRequest.Photo, dishRequest.Id);
 
                 return response.IsSuccessStatusCode;
             }
@@ -158,6 +172,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var json = JsonConvert.SerializeObject(dishRequest);
                 var content = new StringContent(json, Encoding.UTF8, "application/json");
                 var response = await _httpClient.PostAsync($"{_apiUrl}/AddDish", content);
@@ -180,7 +204,7 @@ namespace Luqmit3ish.Services
             }
         }
 
-        internal async Task<bool> UploadPhoto(string photoPath, int foodId)
+        public async Task<bool> UploadPhoto(string photoPath, int foodId)
         {
             if (!_connection.CheckInternetConnection())
             {
@@ -188,6 +212,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 ByteArrayContent fileContent;
                 if (Uri.TryCreate(photoPath, UriKind.Absolute, out Uri uri) && (uri.Scheme == Uri.UriSchemeHttp || uri.Scheme == Uri.UriSchemeHttps))
                 {
@@ -270,6 +304,17 @@ namespace Luqmit3ish.Services
         {
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
+
                 var response = await _httpClient.DeleteAsync($"{_apiUrl}/{food_id}");
 
                 if (!response.IsSuccessStatusCode)
@@ -287,6 +332,6 @@ namespace Luqmit3ish.Services
             }
         }
 
-        
+       
     }
 }
