+++ Luqmit3ish/Luqmit3ish/Services/OrderService.cs
@@ -4,18 +4,21 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.Net;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
 using Luqmit3ish.Connection;
 using Luqmit3ish.Exceptions;
+using Luqmit3ish.Interfaces;
 using Luqmit3ish.Models;
 using Luqmit3ish.Utilities;
 using Luqmit3ish.ViewModels;
 using Newtonsoft.Json;
+using Xamarin.Essentials;
 
 namespace Luqmit3ish.Services
 {
-    class OrderService
+    class OrderService: IOrderService
     {
         private readonly HttpClient _httpClient;
         private  readonly string _apiUrl = Constants.BaseUrl + "api/Orders";
@@ -26,10 +29,10 @@ namespace Luqmit3ish.Services
 
         private IConnection _connection;
 
-            public OrderService()
+        public OrderService()
         {
             _httpClient = new HttpClient();
-            _connection = new Connection();
+            _connection = new InternetConnection();
         }
 
         public async Task<ObservableCollection<OrderCard>> GetOrders(int id)
@@ -40,6 +43,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var response = await _httpClient.GetAsync($"{_orderApiUrl}/{id}");
                 var content = await response.Content.ReadAsStringAsync();
                 return JsonConvert.DeserializeObject<ObservableCollection<OrderCard>>(content);
@@ -56,7 +69,7 @@ namespace Luqmit3ish.Services
 
         }
         
-       public async Task<ObservableCollection<OrderCard>> GetRestaurantOrders(int id,bool receieve)
+        public async Task<ObservableCollection<OrderCard>> GetRestaurantOrders(int id,bool receieve)
         {
             if (!_connection.CheckInternetConnection())
             {
@@ -64,6 +77,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var response = await _httpClient.GetAsync($"{_restaurantApiUrl}/{id}/{receieve}");
                 var content = await response.Content.ReadAsStringAsync();
                 return JsonConvert.DeserializeObject<ObservableCollection<OrderCard>>(content);
@@ -74,8 +97,8 @@ namespace Luqmit3ish.Services
             {
                 throw new Exception(e.Message);
             }
-            
         }
+
         public async Task<DishesOrder> GetBestRestaurant()
         {
             if (!_connection.CheckInternetConnection())
@@ -96,8 +119,8 @@ namespace Luqmit3ish.Services
             {
                 throw new Exception(e.Message);
             }
-
         }
+
         public async Task<Order> GetOrderById(int id)
 
         {
@@ -107,12 +130,19 @@ namespace Luqmit3ish.Services
             }
             try
             {
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var response = await _httpClient.GetAsync($"{_apiUrl}/{id}");
-
-                
-                    var content = await response.Content.ReadAsStringAsync();
-                    return JsonConvert.DeserializeObject<Order>(content);
-                
+                var content = await response.Content.ReadAsStringAsync();
+                return JsonConvert.DeserializeObject<Order>(content);
             }
             catch (HttpRequestException e)
             {
@@ -123,7 +153,6 @@ namespace Luqmit3ish.Services
 
                 throw new Exception($"Failed to retrieve user"+e.Message);
             }
-            
         }
 
         public async Task<bool> UpdateOrderDishCount(int id, string operation)
@@ -134,7 +163,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
-
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var patchObject = new { id, operation };
                 var patchData = JsonConvert.SerializeObject(patchObject);
                 var httpContent = new StringContent(patchData, Encoding.UTF8, "application/json");
@@ -157,10 +195,7 @@ namespace Luqmit3ish.Services
             catch (Exception e)
             {
                 throw new Exception(e.Message);
-
-
             } 
-         
         }
 
         public async Task<bool> ReserveOrder(Order orderRequest)
@@ -168,6 +203,16 @@ namespace Luqmit3ish.Services
             if (!_connection.CheckInternetConnection())
             {
                 throw new ConnectionException("There is no internet connection");
+            }
+            string token = Preferences.Get("Token", string.Empty);
+            if (string.IsNullOrEmpty(token))
+            {
+                _httpClient.DefaultRequestHeaders.Authorization = null;
+                throw new NotAuthorizedException("You are not Authorized to do this operation");
+            }
+            else
+            {
+                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             }
             try
             {
@@ -184,8 +229,9 @@ namespace Luqmit3ish.Services
             {
                 throw new Exception(e.Message);
             }
-
         }
+
+
         public async Task<bool> DeleteOrder(int charityId, int restaurantId)
         {
             if (!_connection.CheckInternetConnection())
@@ -194,7 +240,17 @@ namespace Luqmit3ish.Services
             }
             try
             {
-            var response = await _httpClient.DeleteAsync($"{_apiUrl}/delete/{charityId}/{restaurantId}");
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
+                var response = await _httpClient.DeleteAsync($"{_apiUrl}/delete/{charityId}/{restaurantId}");
                 return response.IsSuccessStatusCode;
             }
             catch (HttpRequestException e)
@@ -207,8 +263,6 @@ namespace Luqmit3ish.Services
                 Debug.WriteLine(e.Message);
                 return false;
             }
-
-
         }
          public async Task<bool> UpdateOrderReceiveStatus(int id)
         {
@@ -218,7 +272,16 @@ namespace Luqmit3ish.Services
             }
             try
             {
-
+                string token = Preferences.Get("Token", string.Empty);
+                if (string.IsNullOrEmpty(token))
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = null;
+                    throw new NotAuthorizedException("You are not Authorized to do this operation");
+                }
+                else
+                {
+                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                }
                 var patchObject = new { id };
                 var patchData = JsonConvert.SerializeObject(patchObject);
                 var httpContent = new StringContent(patchData, Encoding.UTF8, "application/json");
@@ -241,12 +304,7 @@ namespace Luqmit3ish.Services
             catch (Exception e)
             {
                 throw new Exception(e.Message);
-
-
             }
-
         }
     }
 }
-
-
